Coalesce wait_for_condition definitions and set default timeout to 30s
- Merged wait_for_condition from tests/functional/web/ui/conftest.py and tests/helpers.py into a single implementation in tests/helpers.py.
- Increased default timeout to 30.0s.
- Updated implementation to support both sync and async conditions and return the truthy result.
- Updated all call sites and ast-grep rules/documentation to reflect the new signature.
