+++ .ast-grep/rules/no-asyncio-sleep-in-tests.yml
@@ -12,8 +12,8 @@ note: |
     # Wait for a condition to be true
     await wait_for_condition(
         lambda: check_some_state(),
-        timeout_seconds=5.0,
-        poll_interval_seconds=0.1
+        timeout=5.0,
+        interval=0.1
     )
 
   NOTE: This rule allows asyncio.sleep() inside while loops, as those are typically
